# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "weblav",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/backup/restore": {
      "post": {
        "summary": "受け取ったバックアップで、今のデータを置き換える。全員のセッションを消すので、この要求の送り手もログインし直す。",
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreBackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "戻した"
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin以外",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "受け取ったものが無い・期限切れ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "アーカイブの走査中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "バックアップが壊れている",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "ユーザーの一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserListItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin以外",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "作成したユーザー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItem"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin以外",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Free の上限に当たった",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "ユーザー名重複・パスワード未入力",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/setup/admin": {
      "post": {
        "operationId": "create_admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "作成した管理者 (このセッションでログイン済み) とリカバリコード",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithRecoveryCodeResponse"
                }
              }
            }
          },
          "404": {
            "description": "セットアップ済み・トークンが違う・期限切れ・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "ユーザー名重複・ユーザー名/パスワード未入力",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/restore": {
      "post": {
        "summary": "受け取ったバックアップで、今のデータを置き換える。ログインはさせない (戻したデータの管理者でログインし直す)。",
        "operationId": "setup_restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "戻した"
          },
          "404": {
            "description": "セットアップ済み・トークンが違う・期限切れ・受け取ったものが無い・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "アーカイブの走査中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "バックアップが壊れている",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/restore/stage": {
      "post": {
        "summary": "管理者を作る代わりに、バックアップから戻す (→ docs/access.md「バックアップとリストア」)。まず受け取って目録を返す。",
        "operationId": "setup_stage_restore",
        "parameters": [
          {
            "name": "x-setup-token",
            "in": "header",
            "description": "発行したトークン",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/StageBackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "受け取ったバックアップの目録",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagedBackupResponse"
                }
              }
            }
          },
          "404": {
            "description": "セットアップ済み・トークンが違う・期限切れ・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "バックアップではない・新しい版で作られた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/status": {
      "get": {
        "operationId": "status",
        "responses": {
          "200": {
            "description": "管理者がまだいない (セットアップが要る)"
          },
          "404": {
            "description": "セットアップ済み・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/token": {
      "post": {
        "operationId": "issue_token",
        "responses": {
          "200": {
            "description": "発行したトークン",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "セットアップ済み・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup/token/verify": {
      "post": {
        "operationId": "verify_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "使えるトークン"
          },
          "404": {
            "description": "セットアップ済み・トークンが違う・期限切れ・この PC の外からの要求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateAdminRequest": {
        "type": "object",
        "required": [
          "token",
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "role": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Role",
                "description": "権限。省略時は `user`。"
              },
              {
                "type": "null"
              }
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ValidationDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "API エラーの機械可読なコード。frontend はこの値で表示文言を選ぶ。",
        "enum": [
          "not_found",
          "database_unavailable",
          "unauthorized",
          "forbidden",
          "invalid_credentials",
          "incorrect_current_password",
          "invalid_recovery_code",
          "too_many_requests",
          "invalid_request_body",
          "file_too_large",
          "internal_error",
          "conflict",
          "config_unreadable",
          "free_limit_reached",
          "account_server_unavailable",
          "pro_code_invalid",
          "pro_clock_behind"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通envelope。`code` が機械可読な契約(frontendはこちらでi18nする)、\n`message` は英語のデバッグ用の説明文で、画面には出さない。\n\n汎用の文言では直しようがない422には `detail` を添える。こちらも機械可読な契約で、\n画面が値を差し込んで文言を組み立てる (→ `api::error_detail`)。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "RestoreBackupRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "受け取ったときに返した id。"
          }
        }
      },
      "Role": {
        "type": "string",
        "description": "ユーザーの権限。管理者向け機能(コンテンツ管理等)の土台。\n`sqlx::Type` により `users.role` の TEXT カラム(`'admin'`/`'user'`)と相互変換する。",
        "enum": [
          "admin",
          "user"
        ]
      },
...
      "SetupRestoreRequest": {
        "type": "object",
        "required": [
          "token",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "受け取ったときに返した id。"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "StageBackupRequest": {
        "type": "object",
        "description": "utoipa の OpenAPI ドキュメント用のダミー構造体 (実際は `axum::extract::Multipart` で受け取る)。",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "StagedBackupResponse": {
        "type": "object",
        "description": "受け取ったバックアップの目録。確認の画面に出す。",
        "required": [
          "id",
          "createdAt",
          "appVersion"
        ],
        "properties": {
          "appVersion": {
            "type": "string",
            "description": "作った weblav の版。"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "作った日時 (UNIX 秒)。"
          },
          "id": {
            "type": "string",
            "description": "戻すときに指す id。"
          }
        }
      },
...
      "UserListItem": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "createdAt",
          "privateContentCount"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "privateContentCount": {
            "type": "integer",
            "format": "int64",
            "description": "そのユーザーが作成者の `private` の件数。削除の確認で、`hidden` として残る件数を\n示すため (→ docs/access.md「ユーザーの削除と作成者」)。"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "hasRecoveryCode"
        ],
        "properties": {
          "hasRecoveryCode": {
            "type": "boolean",
            "description": "リカバリコードを作ってあるか (→ docs/access.md「リカバリコード」)。管理画面で作るよう促すのに使う。"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserWithRecoveryCodeResponse": {
        "type": "object",
        "description": "利用者と、その人に作ったリカバリコード。初回セットアップとリカバリコードでの\n再設定が返す。コードはこの応答でしか返さない。",
        "required": [
          "user",
          "recoveryCode"
        ],
        "properties": {
          "recoveryCode": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
...
      "VerifyTokenRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  }
}